% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{package}
\name{bitData-package}
\alias{bitData}
\alias{bitData-package}
\title{bitData: 통계(데이터 과학) 데이터 모음집}
\description{
오픈 통계 패키지와 통계(데이터 과학) 한국 데이터 모음
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bit2r.github.io/bitData/}
  \item \url{https://github.com/bit2r/bitData}
  \item Report bugs at \url{https://github.com/bit2r/bitData/issues}
}

}
\author{
\strong{Maintainer}: Kwangchun Lee \email{kwangchun.lee.7@gmail.com}

}
