% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{data}
\name{co2}
\alias{co2}
\title{국가 온실가스 종류별 배출량}
\format{
A data frame with 30 rows and 10 variables:
\describe{
\item{\code{시점}}{integer 시점 1년 단위 연도}
\item{\code{총배출량}}{double 총배출량, 단위: 백만tCO₂}
\item{\code{순배출량}}{double 순배출량, 단위: 백만tCO₂}
\item{\code{CO2}}{double CO2, 이산화탄소}
\item{\code{CH4}}{double CH4, 메테인}
\item{\code{N2O}}{double N2O, 이산화질수}
\item{\code{HFCs}}{double HFCs, (에어컨 냉매) 수소불화탄소}
\item{\code{PFCs}}{character (오존층 파괴) 과불화탄소}
\item{\code{SF6}}{double SF6, 육불화황}
\item{\code{LULUCF}}{double 토지이용, 토지이용 변화 및 임업(LULUCF: Land Use, Land Use Change and Forestry)}
}
}
\source{
\url{https://kosis.kr/statHtml/statHtml.do?orgId=106&tblId=DT_106N_99_2800021&conn_path=I3}
}
\usage{
co2
}
\description{
환경부 온실가스종합정보센터 (국가 온실가스 인벤토리 보고서)에서
}
\details{
저탄소 녹색성장 기본법(제45조)에 따라 매년 국가 온실가스 통계 관리위원회 심의를 거쳐 공표하고 있는 통계표로 KOSIS에서 출처(환경부 온실가스종합정보센터, 국가 온실가스 인벤토리 보고서)를 밝히고 통계표로 제공되고 있다.
}
\keyword{datasets}
