% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{data}
\name{disaster}
\alias{disaster}
\title{대한민국 참사 데이터}
\format{
19개 대형참사에 대한 3가지 변수를 갖는 데이터프레임:
\describe{
\item{\code{날짜}}{date 참사발생일자}
\item{\code{사망}}{double 사망/희생자 수}
\item{\code{참사내역}}{character 참사 명칭} #'}
}
\source{
헤럴드 경제, \link{데이터랩} 역대 대형참사…최악은 502명 사망 삼풍 붕괴 \url{http://news.heraldcorp.com/view.php?ud=20140417000274}

서울경제, 성수대교 붕괴서 이태원 압사까지…잊을만 하면 터지는 '후진국형 대형참사' \url{https://www.sedaily.com/NewsView/26CJW8PC06}
}
\usage{
disaster
}
\description{
대한민국 건국이후 참사에 대한 기록
}
\details{
헤럴드경제 뉴스기사(2014)와 10.29 참사(이태원) 서울경제 실린 참사 데이터
}
\keyword{datasets}
