% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{data}
\name{exam_score}
\alias{exam_score}
\title{영어작문 시험성적}
\format{
수강생 72명에 대한 8개 변수를 갖는 데이터프레임
\describe{
\item{\code{id}}{double 학번}
\item{\code{department}}{character 전공}
\item{\code{year}}{double 대학 학년(1~4)}
\item{\code{attendance}}{double 출석횟수}
\item{\code{participation}}{double 참여횟수, 발표 등}
\item{\code{homework}}{double 과제 점수}
\item{\code{midterm}}{double 중간고사 점수}
\item{\code{final}}{double 기말고사 점수}
}
}
\usage{
exam_score
}
\description{
국내 모대학 영어작문 수강생 시험 평가 점수 데이터셋
}
\details{
Seoul R Meetup 워크샵에서 참여자들이 직접 제출한 데이터
}
\keyword{datasets}
