% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{남극 파머 연구소 인근 펭귄 측정 데이터}
\format{
344개 관측점과 8개 변수를 갖는 데이터프레임:
\describe{
\item{\code{종명칭}}{factor 펭귄종 명칭 ("아델리", "젠투", "턱끈")}
\item{\code{섬이름}}{factor 남극 파머 제도 섬이름 ("드림", "비스코", "토르거센")}
\item{\code{부리_길이}}{double 부리 길이 (밀리미터), 2개 NA 포함}
\item{\code{부리_깊이}}{double 부리 깊이 (밀리미터), 2개 NA 포함}
\item{\code{날개_길이}}{integer 날개 길이 (밀리미터), flipper length, 2개 NA 포함}
\item{\code{체중}}{integer 체중 (그램), 2개 NA 포함}
\item{\code{성별}}{factor 성별 (수컷, 암컷), 11개 NA 포함}
\item{\code{연도}}{integer 측정 연도 (2007, 2008, 2009)}
}
}
\source{
{Horst AM, Hill AP, Gorman KB (2020). palmerpenguins: Palmer Archipelago (Antarctica) penguin data. R package version 0.1.0.} \doi{10.5281/zenodo.3960218}

\url{https://allisonhorst.github.io/palmerpenguins/}

{Adélie penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Adélie penguins (Pygoscelis adeliae) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative.} \doi{10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}

{Gentoo penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Gentoo penguin (Pygoscelis papua) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative.} \doi{10.6073/pasta/7fca67fb28d56ee2ffa3d9370ebda689}

{Chinstrap penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Chinstrap penguin (Pygoscelis antarcticus) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 6. Environmental Data Initiative.} \doi{10.6073/pasta/c14dfcfada8ea13a17536e73eb6fbe9e}

{Originally published in: Gorman KB, Williams TD, Fraser WR (2014) Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis). PLoS ONE 9(3): e90081. doi:10.1371/journal.pone.0090081}
}
\usage{
penguins
}
\description{
남극 파머 연구소 인근 서식하고 있는 성체 펭귄을 관측한 데이터로
펭귄종, 파머 제도 서식섬, 측정크기(날개 길이, 체중, 부리 길이와 깊이), 성별에
대한 정보가 포함되어 있다.
}
\details{
원본 palmerpenguins 패키지의 데이터를 한글로 번역한 버전입니다.
일부 관측치에서 측정값이 누락되어 NA가 포함되어 있습니다 (총 11개 행에 NA 존재).
}
\keyword{datasets}
