% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{data}
\name{population}
\alias{population}
\title{대한민국 인구 데이터}
\format{
관측점으로 총 41년, 변수로 4개를 갖는 데이터프에임:
\describe{
\item{\code{시점}}{integer 시점, 연도 1년 단위}
\item{\code{총인구}}{double 총인구, 단위 (명)}
\item{\code{출생아수}}{double 출생아수, 단위 (명)}
\item{\code{사망자수}}{integer 사망자수, 단위 (명)}
}
}
\source{
\url{https://kosis.kr/statHtml/statHtml.do?orgId=101&tblId=DT_1B8000G&conn_path=I3}

\url{https://kosis.kr/statHtml/statHtml.do?orgId=101&tblId=DT_1BPA002&conn_path=I3}
}
\usage{
population
}
\description{
1981년 이후 연도별 총인구수와 출생아수와 사망자수를 담은 데이터셋
}
\details{
KOSIS로부터 통계청 "장래인구추계" 주요인구지표 및 인구동향조사 통계표를 근간으로 제작됨.
}
\keyword{datasets}
