% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{data}
\name{terms}
\alias{terms}
\title{대한민국 대통령 재임기간}
\format{
대통령 13명에 대한 5개 변수를 갖는 데이터프레임:
\describe{
\item{\code{이름}}{character 대통령 이름}
\item{\code{시작일}}{date 대통령 임기 시작일}
\item{\code{종료일}}{date 대통령 임기 종료일}
\item{\code{임기}}{interval 재임기간}
\item{\code{재임기간}}{double 재임기간 년도숫자 표시}
}
}
\source{
위키백과 대한민국의 대통령 목록 \url{https://ko.wikipedia.org/wiki/대한민국의_대통령_목록}
}
\usage{
terms
}
\description{
역대 대통령 재임기간
}
\details{
해방이후 대한민국 대통령 재임기간을 시작일, 종료일, 임기, 재임기간(년도)로 기록
}
\keyword{datasets}
