% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitData.R
\docType{data}
\name{typhoon}
\alias{typhoon}
\title{자연재해 태풍 희생자}
\format{
10개 태풍에 대한 5개 변수를 갖는 데이터프레임:
\describe{
\item{\code{시작일}}{date 태풍 영향 시작일자}
\item{\code{종료일}}{date 태풍 영향 종료일자}
\item{\code{사망}}{double 태풍으로 인한 인명피해(명)}
\item{\code{태풍명}}{character 태풍명칭}
\item{\code{지속일수}}{double 한반도에 태풍이 영향을 미친 일수}
}
}
\source{
나무위키 태풍/대한민국/각종 기록 순위 \url{https://namu.wiki/w/태풍/대한민국/각종 기록 순위}
}
\usage{
typhoon
}
\description{
대한민국을 강타해서 많은 희생자가 발생한 태풍 기록
}
\details{
한반도를 강타한 대표적인 자연재해 태풍으로 인한 희생자수와 태풍 주요 기록
}
\keyword{datasets}
